pyproject.toml
[tool.poetry]
name = "django-backend-hugo"
version = "0.1.0"
description = "A backend server to help manage content for a hugo blog site"
authors = ["jnutterdev <hello@jnutterdev.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
python-dotenv = "^0.19.2"
django-markdownx = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
